+++ src/Contract/PreservesSafety.php
@@ -1,0 +1,8 @@
+<?php
+
+declare(strict_types=1);
+
+namespace Duon\Boiler\Contract;
+
+/** @api */
+interface PreservesSafety {}